moved nine patch obtain logic from nine patch into asset repository
