+++ editor/src/com/talosvfx/talos/editor/addons/scene/MainRenderer.java
@@ -15,6 +15,7 @@ import com.badlogic.gdx.utils.Array;
 import com.badlogic.gdx.utils.ObjectMap;
 import com.esotericsoftware.spine.TalosSkeletonRenderer;
 import com.talosvfx.talos.editor.addons.scene.assets.AssetRepository;
+import com.talosvfx.talos.runtime.RuntimeContext;
 import com.talosvfx.talos.runtime.assets.GameAsset;
 import com.talosvfx.talos.runtime.assets.GameAssetType;
 import com.talosvfx.talos.runtime.assets.GameResourceOwner;
@@ -73,7 +74,6 @@ public class MainRenderer implements Observer {
     private static final int RT = 2;
     private static final int RB = 3;
 
-    private ObjectMap<GameAsset<Texture>, NinePatch> patchCache = new ObjectMap<>();
     private ObjectMap<ParticleComponent, ParticleEffectInstance> particleCache = new ObjectMap<>();
 
     private SpriteBatchParticleRenderer talosRenderer;
@@ -467,7 +467,7 @@ public class MainRenderer implements Observer {
                 final float height = spriteRenderer.size.y;
 
                 if(metadata != null && metadata.borderData != null && spriteRenderer.renderMode == SpriteRendererComponent.RenderMode.sliced) {
-                    NinePatch patch = obtainNinePatch(gameResource);// todo: this has to be done better
+                    NinePatch patch = RuntimeContext.getInstance().AssetRepository.obtainNinePatch(gameResource);// todo: this has to be done better
                     //todo: and this renders wrong so this needs fixing too
                     float xSign = width < 0 ? -1 : 1;
                     float ySign = height < 0 ? -1 : 1;
@@ -646,18 +646,6 @@ public class MainRenderer implements Observer {
         }
     }
 
-    public NinePatch obtainNinePatch (GameAsset<Texture> gameAsset) {
-        if (patchCache.containsKey(gameAsset)) { //something better, maybe hash on pixel size + texture for this
-            return patchCache.get(gameAsset);
-        } else {
-            final SpriteMetadata metadata = (SpriteMetadata) gameAsset.getRootRawAsset().metaData;
-            final NinePatch patch = new NinePatch(gameAsset.getResource(), metadata.borderData[0], metadata.borderData[1], metadata.borderData[2], metadata.borderData[3]);
-            patch.scale(1 / metadata.pixelsPerUnit, 1 / metadata.pixelsPerUnit); // fix this later
-            patchCache.put(gameAsset, patch);
-            return patch;
-        }
-    }
-
     private ParticleEffectInstance obtainParticle (GameObject gameObject, ParticleEffectDescriptor descriptor) {
         ParticleComponent component = gameObject.getComponent(ParticleComponent.class);
 
@@ -747,20 +735,5 @@ public class MainRenderer implements Observer {
 
     public Camera getCamera() {
         return camera;
-    }
-
-
-    @EventHandler
-    public void onSpritePixelPerUnitUpdateEvent (SpritePixelPerUnitUpdateEvent event) {
-        final SpriteMetadata metadata = event.getSpriteMetadata();
-        for (ObjectMap.Entry<GameAsset<Texture>, NinePatch> gameAssetNinePatchEntry : patchCache) {
-            if (gameAssetNinePatchEntry.key.getRootRawAsset().metaData.equals(metadata)) {
-                final NinePatch patch = new NinePatch(gameAssetNinePatchEntry.key.getResource(), metadata.borderData[0], metadata.borderData[1], metadata.borderData[2], metadata.borderData[3]);
-                final float scale = 1 / metadata.pixelsPerUnit;
-                patch.scale(scale, scale);
-                patchCache.put(gameAssetNinePatchEntry.key, patch);
-                break;
-            }
-        }
     }
 }
