+++ editor/src/com/talosvfx/talos/editor/addons/scene/assets/AssetRepository.java
@@ -27,10 +27,7 @@ import com.badlogic.gdx.utils.reflect.ClassReflection;
 import com.badlogic.gdx.utils.reflect.ReflectionException;
 import com.esotericsoftware.spine.SkeletonBinary;
 import com.esotericsoftware.spine.SkeletonData;
-import com.talosvfx.talos.editor.addons.scene.events.AssetColorFillEvent;
-import com.talosvfx.talos.editor.addons.scene.events.AssetPathChanged;
-import com.talosvfx.talos.editor.addons.scene.events.AssetResolutionChanged;
-import com.talosvfx.talos.editor.addons.scene.events.ScriptFileChangedEvent;
+import com.talosvfx.talos.editor.addons.scene.events.*;
 import com.talosvfx.talos.editor.addons.scene.events.meta.MetaDataReloadedEvent;
 import com.talosvfx.talos.runtime.assets.AMetadata;
 import com.talosvfx.talos.editor.addons.scene.utils.importers.AssetImporter;
@@ -48,6 +45,7 @@ import com.talosvfx.talos.runtime.assets.BaseAssetRepository;
 import com.talosvfx.talos.runtime.assets.meta.DirectoryMetadata;
 import com.talosvfx.talos.runtime.assets.meta.ScriptMetadata;
 import com.talosvfx.talos.runtime.assets.meta.SpineMetadata;
+import com.talosvfx.talos.runtime.assets.meta.SpriteMetadata;
 import com.talosvfx.talos.runtime.maps.TilePaletteData;
 import com.talosvfx.talos.runtime.utils.NamingUtils;
 import com.talosvfx.talos.editor.utils.Toasts;
@@ -88,7 +86,7 @@ public class AssetRepository extends BaseAssetRepository implements Observer {
 	public static final AssetNameFieldFilter ASSET_NAME_FIELD_FILTER = new AssetNameFieldFilter();
 
 	private ObjectMap<GameAssetType, ObjectMap<String, GameAsset<?>>> identifierGameAssetMap = new ObjectMap<>();
-
+	private ObjectMap<GameAsset<Texture>, NinePatch> patchCache = new ObjectMap<>();
 	private ObjectSet<FileHandle> newFilesSeen = new ObjectSet<>();
 
 	public <T> GameAsset<T> getAssetForUniqueIdentifier (String uuid, GameAssetType type) {
@@ -118,7 +116,29 @@ public class AssetRepository extends BaseAssetRepository implements Observer {
 
 	@Override
 	public NinePatch obtainNinePatch (GameAsset<Texture> gameAsset) {
-		return null;
+		if (patchCache.containsKey(gameAsset)) { //something better, maybe hash on pixel size + texture for this
+			return patchCache.get(gameAsset);
+		} else {
+			final SpriteMetadata metadata = (SpriteMetadata) gameAsset.getRootRawAsset().metaData;
+			final NinePatch patch = new NinePatch(gameAsset.getResource(), metadata.borderData[0], metadata.borderData[1], metadata.borderData[2], metadata.borderData[3]);
+			patch.scale(1 / metadata.pixelsPerUnit, 1 / metadata.pixelsPerUnit); // fix this later
+			patchCache.put(gameAsset, patch);
+			return patch;
+		}
+	}
+
+	@EventHandler
+	public void onSpritePixelPerUnitUpdateEvent (SpritePixelPerUnitUpdateEvent event) {
+		final SpriteMetadata metadata = event.getSpriteMetadata();
+		for (ObjectMap.Entry<GameAsset<Texture>, NinePatch> gameAssetNinePatchEntry : patchCache) {
+			if (gameAssetNinePatchEntry.key.getRootRawAsset().metaData.equals(metadata)) {
+				final NinePatch patch = new NinePatch(gameAssetNinePatchEntry.key.getResource(), metadata.borderData[0], metadata.borderData[1], metadata.borderData[2], metadata.borderData[3]);
+				final float scale = 1 / metadata.pixelsPerUnit;
+				patch.scale(scale, scale);
+				patchCache.put(gameAssetNinePatchEntry.key, patch);
+				break;
+			}
+		}
 	}
 
 	private <T> void putAssetForIdentifier (String identifier, GameAssetType type, GameAsset<T> asset) {
